Changed houses list and action card list to fit the format cards.java
player.houses[] changed from list of cards to list of HouseCards.
Deleted player.ActionCards list and replaced it with an int
numberActionCards which tracks the number of Action Cards collected.
Added method getNumberActionCards() to retrieve number of ActionCards
and added method increaseActionCards() which increments the number of
ActionCards held by the player.