Added Children Enum, Spinner Class and sellHouse() method in Player.java
Children Enum to avoid number of children exceeding the limit.

Added Spinner Class, with method spin().

Added sellHouse() method to Player class. It adds the sale price of the
house to the player balance, it then removes the house from that players
list. Returns the HouseCard so it can be added back to the deck.